mgr/dashboard：delete image fail on dashboard while it has some protected snapshots Fixes:  https://tracker.ceph.com/issues/62047 Signed-off-by: Teng Jie <tengjie5@asiainfo.com>

+++ src/pybind/mgr/dashboard/services/rbd.py
@@ -3,6 +3,7 @@
 import errno
 import json
 import math
+import time
 from enum import IntEnum
 
 import cherrypy
@@ -630,9 +631,16 @@ class RbdService(object):
 
         image = RbdService.get_image(image_spec)
         snapshots = image['snapshots']
-        for snap in snapshots:
-            RbdSnapshotService.remove_snapshot(image_spec, snap['name'], snap['is_protected'])
-
+        if len(snapshots) > 0:
+            count = 0
+            for snap in snapshots:
+                RbdSnapshotService.remove_snapshot(image_spec, snap['name'], snap['is_protected'])
+            while True:
+                image = RbdService.get_image(image_spec)
+                if len(image['snapshots']) < 1 or count > 30:
+                    break
+                time.sleep(1)
+                count += 1
         rbd_inst = rbd.RBD()
         return rbd_call(pool_name, namespace, rbd_inst.remove, image_name)
